Get starships from API
Here we wire up the API to get the starships from: `https://swapi.dev/api/starships/`.

We do manually do the parsing of the JSON into our `Starship` model.

We user `Alamofire` to make the simple network call and `SwiftyJSON` to parse the result.

IMPORTANT: in a real life situation we most likely wouldn't be using `SwiftyJSON` and instead the model objects and the parsing code itself would be automatically be generated from an `OpenAPI` doc.