Show starships in a list
Here we take the retrieved `Starship`s and show them in a collection view. We use a collection view as opposed to a table view as it's more future proof e.g. it's plausible that the next stage of development might be to show the starships in a carousel.

We show the following properties in the list:

* Name
* Manufacturer
* Cost (if available)

And allow the user to tap a "star" button to favourite a starship. We store the favourites (in memory) in the `AppManager`. To change to persistable is simple as making the `Starship` codable and writing custom getters and setters for `favouritedStarships` to write to `userDefaults`.

The `Starship`s are sortable by tapping the sort button in the top right. We've chosen a random subset of `Starship` properties as supported sort options for demonstration purposes.

The business logic (loading, favouriting, sorting etc.) is done in the presenter while the display logic is done in the view controller. NB: when a user taps a favourite button, we have a handler in the cell that calls the delegate function `didToggleFavourite` -- the view controller conforms to `StarshipListCollectionViewCellDelegate`, but simply passes the call to the presenter to handle the business logic.

NB: due to time constraints there's various critical things that I dropped -- I've noted these as `TODO`s.

We've included another 3rd party library: `ActionSheetPicker-3.0` to use for our sort order selection.