Starship details
Here we make a screen to show the starship details when the user taps on a starship from the main list.

The presenter calls the router which then configures the details scene.

In the details screen we have the same basic layout as the cells in the main list, but we simply show more details.

NB: there is a room to add much more fields on the screen, but I'm guessing that fully fleshing out this screen is not the main point of the exercise :-).

NB: I've noted with `TODO`s the stuff I would refactor if it were a real life project.